Add additional checks of visibility/availability state.
This PR adds additional checks and a new setting to
control the way the block handles coursemodulesvisibiity
and availability.

The properties to check are: visible, visibleoncoursepage,
uservisible and available.

The first two are static properties from the coursemodules
table. However the last ones are dynamic props that may be
influenced by a explicit setting or role permission.

Fo that reason it could be more close to the "moodle way"
to structure the logic, prioritizing the static properties
and validating possible reasons(influences) that affects
the current availability/visibility of the dynamic props.

An example: if an activity is hidden, it may be still
accessed by the student if the admin allows that via
role permissions.

Another example: the activity may be restricted to
students that don't met criteria. However, by default
the course page allows to see the link in a inactive
state.

For this last one example, the admin may decide the behaviour
via a new plugin admin setting "toggleshowrestricted".